requirements.txt
alembic==1.4.1
backports.entry-points-selectable==1.1.0
black==21.9b0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
configparser==5.0.2
databricks-cli==0.15.0
distlib==0.3.3
docker==5.0.3
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.3.1
Flask==2.0.2
gitdb==4.0.7
GitPython==3.1.24
greenlet==1.1.2
gunicorn==20.1.0
identify==2.3.0
idna==3.3
importlib-metadata==4.8.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
Mako==1.1.5
MarkupSafe==2.0.1
mlflow==1.20.2
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.21.2
packaging==21.0
pandas==1.3.4
pathspec==0.9.0
pathtools==0.1.2
Pillow==8.4.0
platformdirs==2.4.0
pre-commit==2.15.0
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.3
promise==2.3
protobuf==3.18.1
psutil==5.8.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.3
PyYAML==6.0
querystring-parser==1.2.4
regex==2021.10.8
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
sentry-sdk==1.4.3
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.25
sqlparse==0.4.2
subprocess32==3.5.4
tabulate==0.8.9
termcolor==1.1.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.1
torch==1.9.1
torchvision==0.10.1
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.8.1
wandb==0.12.4
websocket-client==1.2.1
Werkzeug==2.0.2
yaspin==2.1.0
zipp==3.6.0
